requirements.txt
alembic==1.12.0
blinker==1.6.2
CacheControl==0.13.1
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
dnspython==2.4.2
email-validator==2.1.0.post1
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
mysql-connector-python==8.1.0
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
WTForms==3.0.1
